requirements.txt
amqp==2.5.2
anyjson==0.3.3
billiard==3.5.0.5
celery==4.2.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.11
django-background-tasks==1.2.5
django-celery-beat==2.0.0
django-compat==1.0.15
django-timezone-field==4.0
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-crontab==2.4.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.2.0
requests==2.23.0
schedule==0.6.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
zipp==3.1.0
